+++ src/ni_measurement_plugin_converter/constants/_measurement_service.py
@@ -1,5 +1,20 @@
 """Constants utilized in measurement file."""
 
+from enum import Enum
+
+
+class DriverSession(Enum):
+    """Instrument drivers' session."""
+
+    nidcpower = "nims.session_management.INSTRUMENT_TYPE_NI_DCPOWER"
+    nidmm = "nims.session_management.INSTRUMENT_TYPE_NI_DMM"
+    nidigital = "nims.session_management.INSTRUMENT_TYPE_NI_DIGITAL_PATTERN"
+    nifgen = "nims.session_management.INSTRUMENT_TYPE_NI_FGEN"
+    niscope = "nims.session_management.INSTRUMENT_TYPE_NI_SCOPE"
+    niswitch = "nims.session_management.INSTRUMENT_TYPE_NI_RELAY_DRIVER"
+    nidaqmx = "nims.session_management.INSTRUMENT_TYPE_NI_DAQMX"
+
+
 # Python native data types and its corresponding `measurement_plugin_sdk_service` data types.
 NIMS_TYPE = {
     "int": "nims.DataType.Int64",