+++ src/ni_measurement_plugin_converter/models/_sessions.py
@@ -1,0 +1,10 @@
+"""Models utilized in session management."""
+
+from pydantic import BaseModel
+
+
+class SessionMapping(BaseModel):
+    """Session mapping model."""
+
+    name: str
+    mapping: str